积分优化

+++ src/main/java/com/wanhutong/backend/modules/biz/dao/integration/BizMoneyRecodeDao.java
@@ -10,6 +10,7 @@ import com.wanhutong.backend.modules.biz.entity.integration.BizMoneyRecodeDetail
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -29,6 +30,6 @@ public interface BizMoneyRecodeDao extends CrudDao<BizMoneyRecode> {
     void updateMoney(List<BizMoneyRecode> list);
 
     @Select("SELECT money from biz_cust_credit where office_id = #{officeId}")
-    Double selectMoney(@Param("officeId") Integer officeId);
+    BigDecimal selectMoney(@Param("officeId") Integer officeId);
 	
 }
